+++ packages/sdks/nextjs-sdk/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.8](https://github.com/descope/descope-js/compare/nextjs-sdk-0.2.7...nextjs-sdk-0.2.8) (2024-09-12)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.0.74`
+## [0.2.7](https://github.com/descope/descope-js/compare/nextjs-sdk-0.2.6...nextjs-sdk-0.2.7) (2024-09-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.16.5`
+* `react-sdk` updated to version `2.0.73`
+* `core-js-sdk` updated to version `2.25.0`
+* `web-component` updated to version `3.24.2`
+## [0.2.6](https://github.com/descope/descope-js/compare/nextjs-sdk-0.2.5...nextjs-sdk-0.2.6) (2024-09-05)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.0.72`
 ## [0.2.5](https://github.com/descope/descope-js/compare/nextjs-sdk-0.2.4...nextjs-sdk-0.2.5) (2024-09-03)
 
 ### Dependency Updates