+++ packages/sdks/vue-sdk/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.0.33](https://github.com/descope/descope-js/compare/vue-sdk-2.0.32...vue-sdk-2.0.33) (2024-09-11)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.121`
+* `audit-management-widget` updated to version `0.1.85`
+* `role-management-widget` updated to version `0.1.119`
+* `user-management-widget` updated to version `0.4.122`
+* `user-profile-widget` updated to version `0.0.103`
+* `web-component` updated to version `3.24.2`
+* `web-js-sdk` updated to version `1.16.5`
+* `core-js-sdk` updated to version `2.25.0`
+## [2.0.32](https://github.com/descope/descope-js/compare/vue-sdk-2.0.31...vue-sdk-2.0.32) (2024-09-05)
+
+### Dependency Updates
+
+* `audit-management-widget` updated to version `0.1.84`
 ## [2.0.31](https://github.com/descope/descope-js/compare/vue-sdk-2.0.30...vue-sdk-2.0.31) (2024-09-03)
 
 ### Dependency Updates